requirements.txt
asgiref==3.2.3
Django==2.1.15
django-filter==2.2.0
django-pyodbc-azure==2.1.0.0
djangorestframework==3.11.0
Pillow==6.2.1
pyodbc==4.0.27
pytz==2019.3
sqlparse==0.3.0
XlsxWriter==1.2.6
xlwt==1.3.0
